ISAICP-2369: Refactor the screenshot taking code.
Remove D7 specific code, and change it from a trait to a Context class, this
doesn't make sense as a trait - since it contains step definitions it is
not possible to include it in more than one class.
