+++ tests/features/bootstrap/FeatureContext.php
@@ -19,8 +19,6 @@ class FeatureContext extends RawDrupalContext implements SnippetAcceptingContext
   use ContextualLinksTrait;
   use EntityTrait;
   use UtilityTrait;
-  // Uncomment to enable screen-shots during development.
-  // use \Drupal\joinup\Traits\ScreenShotTrait;
 
   /**
    * Define ASCII values for key presses.
